Deleted ManageNeuronResponse from governance.proto. Seems we are unable to delete ManageNeuron though, because it is used by ProposalData, which is DEFINITELY stored in stable memory.
