+++ rs/nns/governance/src/gen/ic_nns_governance.pb.v1.rs
@@ -1462,257 +1462,6 @@ pub mod manage_neuron {
         RefreshVotingPower(RefreshVotingPower),
     }
 }
-/// The response of the ManageNeuron command
-///
-/// There is a dedicated response type for each `ManageNeuron.command` field
-#[derive(
-    candid::CandidType,
-    candid::Deserialize,
-    serde::Serialize,
-    comparable::Comparable,
-    Clone,
-    PartialEq,
-    ::prost::Message,
-)]
-pub struct ManageNeuronResponse {
-    #[prost(
-        oneof = "manage_neuron_response::Command",
-        tags = "1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14"
-    )]
-    pub command: ::core::option::Option<manage_neuron_response::Command>,
-}
-/// Nested message and enum types in `ManageNeuronResponse`.
-pub mod manage_neuron_response {
-    #[derive(
-        candid::CandidType,
-        candid::Deserialize,
-        serde::Serialize,
-        comparable::Comparable,
-        Clone,
-        Copy,
-        PartialEq,
-        ::prost::Message,
-    )]
-    pub struct ConfigureResponse {}
-    #[derive(
-        candid::CandidType,
-        candid::Deserialize,
-        serde::Serialize,
-        comparable::Comparable,
-        Clone,
-        Copy,
-        PartialEq,
-        ::prost::Message,
-    )]
-    pub struct DisburseResponse {
-        /// The block height at which the disburse transfer happened
-        #[prost(uint64, tag = "1")]
-        pub transfer_block_height: u64,
-    }
-    #[derive(
-        candid::CandidType,
-        candid::Deserialize,
-        serde::Serialize,
-        comparable::Comparable,
-        Clone,
-        Copy,
-        PartialEq,
-        ::prost::Message,
-    )]
-    pub struct SpawnResponse {
-        /// The ID of the Neuron created from spawning a Neuron
-        #[prost(message, optional, tag = "1")]
-        pub created_neuron_id: ::core::option::Option<::ic_nns_common::pb::v1::NeuronId>,
-    }
-    #[derive(
-        candid::CandidType,
-        candid::Deserialize,
-        serde::Serialize,
-        comparable::Comparable,
-        Clone,
-        Copy,
-        PartialEq,
-        ::prost::Message,
-    )]
-    pub struct MergeMaturityResponse {
-        #[prost(uint64, tag = "1")]
-        pub merged_maturity_e8s: u64,
-        #[prost(uint64, tag = "2")]
-        pub new_stake_e8s: u64,
-    }
-    #[derive(
-        candid::CandidType,
-        candid::Deserialize,
-        serde::Serialize,
-        comparable::Comparable,
-        Clone,
-        Copy,
-        PartialEq,
-        ::prost::Message,
-    )]
-    pub struct StakeMaturityResponse {
-        #[prost(uint64, tag = "1")]
-        pub maturity_e8s: u64,
-        #[prost(uint64, tag = "2")]
-        pub staked_maturity_e8s: u64,
-    }
-    #[derive(
-        candid::CandidType,
-        candid::Deserialize,
-        serde::Serialize,
-        comparable::Comparable,
-        Clone,
-        Copy,
-        PartialEq,
-        ::prost::Message,
-    )]
-    pub struct FollowResponse {}
-    #[derive(
-        candid::CandidType,
-        candid::Deserialize,
-        serde::Serialize,
-        comparable::Comparable,
-        Clone,
-        PartialEq,
-        ::prost::Message,
-    )]
-    pub struct MakeProposalResponse {
-        /// The ID of the created proposal
-        #[prost(message, optional, tag = "1")]
-        pub proposal_id: ::core::option::Option<::ic_nns_common::pb::v1::ProposalId>,
-        #[prost(string, optional, tag = "2")]
-        pub message: ::core::option::Option<::prost::alloc::string::String>,
-    }
-    #[derive(
-        candid::CandidType,
-        candid::Deserialize,
-        serde::Serialize,
-        comparable::Comparable,
-        Clone,
-        Copy,
-        PartialEq,
-        ::prost::Message,
-    )]
-    pub struct RegisterVoteResponse {}
-    #[derive(
-        candid::CandidType,
-        candid::Deserialize,
-        serde::Serialize,
-        comparable::Comparable,
-        Clone,
-        Copy,
-        PartialEq,
-        ::prost::Message,
-    )]
-    pub struct SplitResponse {
-        /// The ID of the Neuron created from splitting another Neuron
-        #[prost(message, optional, tag = "1")]
-        pub created_neuron_id: ::core::option::Option<::ic_nns_common::pb::v1::NeuronId>,
-    }
-    /// A response for merging or simulating merge neurons
-    #[derive(
-        candid::CandidType,
-        candid::Deserialize,
-        serde::Serialize,
-        comparable::Comparable,
-        Clone,
-        PartialEq,
-        ::prost::Message,
-    )]
-    pub struct MergeResponse {
-        /// The resulting state of the source neuron
-        #[prost(message, optional, tag = "1")]
-        pub source_neuron: ::core::option::Option<super::Neuron>,
-        /// The resulting state of the target neuron
-        #[prost(message, optional, tag = "2")]
-        pub target_neuron: ::core::option::Option<super::Neuron>,
-        /// The NeuronInfo of the source neuron
-        #[prost(message, optional, tag = "3")]
-        pub source_neuron_info: ::core::option::Option<super::NeuronInfo>,
-        /// The NeuronInfo of the target neuron
-        #[prost(message, optional, tag = "4")]
-        pub target_neuron_info: ::core::option::Option<super::NeuronInfo>,
-    }
-    #[derive(
-        candid::CandidType,
-        candid::Deserialize,
-        serde::Serialize,
-        comparable::Comparable,
-        Clone,
-        Copy,
-        PartialEq,
-        ::prost::Message,
-    )]
-    pub struct DisburseToNeuronResponse {
-        /// The ID of the Neuron created from disbursing a Neuron
-        #[prost(message, optional, tag = "1")]
-        pub created_neuron_id: ::core::option::Option<::ic_nns_common::pb::v1::NeuronId>,
-    }
-    #[derive(
-        candid::CandidType,
-        candid::Deserialize,
-        serde::Serialize,
-        comparable::Comparable,
-        Clone,
-        Copy,
-        PartialEq,
-        ::prost::Message,
-    )]
-    pub struct ClaimOrRefreshResponse {
-        #[prost(message, optional, tag = "1")]
-        pub refreshed_neuron_id: ::core::option::Option<::ic_nns_common::pb::v1::NeuronId>,
-    }
-    #[derive(
-        candid::CandidType,
-        candid::Deserialize,
-        serde::Serialize,
-        comparable::Comparable,
-        Clone,
-        Copy,
-        PartialEq,
-        ::prost::Message,
-    )]
-    pub struct RefreshVotingPowerResponse {}
-    #[derive(
-        candid::CandidType,
-        candid::Deserialize,
-        serde::Serialize,
-        comparable::Comparable,
-        Clone,
-        PartialEq,
-        ::prost::Oneof,
-    )]
-    pub enum Command {
-        #[prost(message, tag = "1")]
-        Error(super::GovernanceError),
-        #[prost(message, tag = "2")]
-        Configure(ConfigureResponse),
-        #[prost(message, tag = "3")]
-        Disburse(DisburseResponse),
-        #[prost(message, tag = "4")]
-        Spawn(SpawnResponse),
-        #[prost(message, tag = "5")]
-        Follow(FollowResponse),
-        #[prost(message, tag = "6")]
-        MakeProposal(MakeProposalResponse),
-        #[prost(message, tag = "7")]
-        RegisterVote(RegisterVoteResponse),
-        #[prost(message, tag = "8")]
-        Split(SplitResponse),
-        #[prost(message, tag = "9")]
-        DisburseToNeuron(DisburseToNeuronResponse),
-        #[prost(message, tag = "10")]
-        ClaimOrRefresh(ClaimOrRefreshResponse),
-        #[prost(message, tag = "11")]
-        MergeMaturity(MergeMaturityResponse),
-        #[prost(message, tag = "12")]
-        Merge(MergeResponse),
-        #[prost(message, tag = "13")]
-        StakeMaturity(StakeMaturityResponse),
-        #[prost(message, tag = "14")]
-        RefreshVotingPower(RefreshVotingPowerResponse),
-    }
-}
 #[derive(candid::CandidType, candid::Deserialize, serde::Serialize, comparable::Comparable)]
 #[compare_default]
 #[derive(Clone, PartialEq, ::prost::Message)]
