+++ rs/nns/governance/src/governance/merge_neurons.rs
@@ -5,12 +5,12 @@ use crate::{
     pb::v1::{
         governance_error::ErrorType,
         manage_neuron::{Merge, NeuronIdOrSubaccount},
-        manage_neuron_response::MergeResponse,
         GovernanceError, NeuronState, ProposalData, ProposalStatus, VotingPowerEconomics,
     },
 };
 use ic_base_types::PrincipalId;
 use ic_nns_common::pb::v1::NeuronId;
+use ic_nns_governance_api::pb::v1::{self as api, manage_neuron_response::MergeResponse};
 use icp_ledger::Subaccount;
 use std::collections::BTreeMap;
 
@@ -349,20 +349,28 @@ pub fn build_merge_neurons_response(
     now_seconds: u64,
     requester: PrincipalId,
 ) -> MergeResponse {
-    let source_neuron = Some(
+    let source_neuron = Some(api::Neuron::from(
         source
             .clone()
             .into_proto(voting_power_economics, now_seconds),
-    );
-    let target_neuron = Some(
+    ));
+    let target_neuron = Some(api::Neuron::from(
         target
             .clone()
             .into_proto(voting_power_economics, now_seconds),
-    );
-    let source_neuron_info =
-        Some(source.get_neuron_info(voting_power_economics, now_seconds, requester));
-    let target_neuron_info =
-        Some(target.get_neuron_info(voting_power_economics, now_seconds, requester));
+    ));
+
+    let source_neuron_info = Some(api::NeuronInfo::from(source.get_neuron_info(
+        voting_power_economics,
+        now_seconds,
+        requester,
+    )));
+    let target_neuron_info = Some(api::NeuronInfo::from(target.get_neuron_info(
+        voting_power_economics,
+        now_seconds,
+        requester,
+    )));
+
     MergeResponse {
         source_neuron,
         target_neuron,
