+++ rs/nns/governance/src/pb/conversions.rs
@@ -1316,361 +1316,6 @@ impl From<pb_api::ManageNeuronCommandRequest> for pb::manage_neuron::Command {
     }
 }
 
-impl From<pb::ManageNeuronResponse> for pb_api::ManageNeuronResponse {
-    fn from(item: pb::ManageNeuronResponse) -> Self {
-        Self {
-            command: item.command.map(|x| x.into()),
-        }
-    }
-}
-impl From<pb_api::ManageNeuronResponse> for pb::ManageNeuronResponse {
-    fn from(item: pb_api::ManageNeuronResponse) -> Self {
-        Self {
-            command: item.command.map(|x| x.into()),
-        }
-    }
-}
-
-impl From<pb::manage_neuron_response::ConfigureResponse>
-    for pb_api::manage_neuron_response::ConfigureResponse
-{
-    fn from(_: pb::manage_neuron_response::ConfigureResponse) -> Self {
-        Self {}
-    }
-}
-impl From<pb_api::manage_neuron_response::ConfigureResponse>
-    for pb::manage_neuron_response::ConfigureResponse
-{
-    fn from(_: pb_api::manage_neuron_response::ConfigureResponse) -> Self {
-        Self {}
-    }
-}
-
-impl From<pb::manage_neuron_response::DisburseResponse>
-    for pb_api::manage_neuron_response::DisburseResponse
-{
-    fn from(item: pb::manage_neuron_response::DisburseResponse) -> Self {
-        Self {
-            transfer_block_height: item.transfer_block_height,
-        }
-    }
-}
-impl From<pb_api::manage_neuron_response::DisburseResponse>
-    for pb::manage_neuron_response::DisburseResponse
-{
-    fn from(item: pb_api::manage_neuron_response::DisburseResponse) -> Self {
-        Self {
-            transfer_block_height: item.transfer_block_height,
-        }
-    }
-}
-
-impl From<pb::manage_neuron_response::SpawnResponse>
-    for pb_api::manage_neuron_response::SpawnResponse
-{
-    fn from(item: pb::manage_neuron_response::SpawnResponse) -> Self {
-        Self {
-            created_neuron_id: item.created_neuron_id,
-        }
-    }
-}
-impl From<pb_api::manage_neuron_response::SpawnResponse>
-    for pb::manage_neuron_response::SpawnResponse
-{
-    fn from(item: pb_api::manage_neuron_response::SpawnResponse) -> Self {
-        Self {
-            created_neuron_id: item.created_neuron_id,
-        }
-    }
-}
-
-impl From<pb::manage_neuron_response::MergeMaturityResponse>
-    for pb_api::manage_neuron_response::MergeMaturityResponse
-{
-    fn from(item: pb::manage_neuron_response::MergeMaturityResponse) -> Self {
-        Self {
-            merged_maturity_e8s: item.merged_maturity_e8s,
-            new_stake_e8s: item.new_stake_e8s,
-        }
-    }
-}
-impl From<pb_api::manage_neuron_response::MergeMaturityResponse>
-    for pb::manage_neuron_response::MergeMaturityResponse
-{
-    fn from(item: pb_api::manage_neuron_response::MergeMaturityResponse) -> Self {
-        Self {
-            merged_maturity_e8s: item.merged_maturity_e8s,
-            new_stake_e8s: item.new_stake_e8s,
-        }
-    }
-}
-
-impl From<pb::manage_neuron_response::StakeMaturityResponse>
-    for pb_api::manage_neuron_response::StakeMaturityResponse
-{
-    fn from(item: pb::manage_neuron_response::StakeMaturityResponse) -> Self {
-        Self {
-            maturity_e8s: item.maturity_e8s,
-            staked_maturity_e8s: item.staked_maturity_e8s,
-        }
-    }
-}
-impl From<pb_api::manage_neuron_response::StakeMaturityResponse>
-    for pb::manage_neuron_response::StakeMaturityResponse
-{
-    fn from(item: pb_api::manage_neuron_response::StakeMaturityResponse) -> Self {
-        Self {
-            maturity_e8s: item.maturity_e8s,
-            staked_maturity_e8s: item.staked_maturity_e8s,
-        }
-    }
-}
-
-impl From<pb::manage_neuron_response::RefreshVotingPowerResponse>
-    for pb_api::manage_neuron_response::RefreshVotingPowerResponse
-{
-    fn from(_item: pb::manage_neuron_response::RefreshVotingPowerResponse) -> Self {
-        Self {}
-    }
-}
-impl From<pb_api::manage_neuron_response::RefreshVotingPowerResponse>
-    for pb::manage_neuron_response::RefreshVotingPowerResponse
-{
-    fn from(_item: pb_api::manage_neuron_response::RefreshVotingPowerResponse) -> Self {
-        Self {}
-    }
-}
-
-impl From<pb::manage_neuron_response::FollowResponse>
-    for pb_api::manage_neuron_response::FollowResponse
-{
-    fn from(_: pb::manage_neuron_response::FollowResponse) -> Self {
-        Self {}
-    }
-}
-impl From<pb_api::manage_neuron_response::FollowResponse>
-    for pb::manage_neuron_response::FollowResponse
-{
-    fn from(_: pb_api::manage_neuron_response::FollowResponse) -> Self {
-        Self {}
-    }
-}
-
-impl From<pb::manage_neuron_response::MakeProposalResponse>
-    for pb_api::manage_neuron_response::MakeProposalResponse
-{
-    fn from(item: pb::manage_neuron_response::MakeProposalResponse) -> Self {
-        Self {
-            proposal_id: item.proposal_id,
-            message: item.message,
-        }
-    }
-}
-impl From<pb_api::manage_neuron_response::MakeProposalResponse>
-    for pb::manage_neuron_response::MakeProposalResponse
-{
-    fn from(item: pb_api::manage_neuron_response::MakeProposalResponse) -> Self {
-        Self {
-            proposal_id: item.proposal_id,
-            message: item.message,
-        }
-    }
-}
-
-impl From<pb::manage_neuron_response::RegisterVoteResponse>
-    for pb_api::manage_neuron_response::RegisterVoteResponse
-{
-    fn from(_: pb::manage_neuron_response::RegisterVoteResponse) -> Self {
-        Self {}
-    }
-}
-impl From<pb_api::manage_neuron_response::RegisterVoteResponse>
-    for pb::manage_neuron_response::RegisterVoteResponse
-{
-    fn from(_: pb_api::manage_neuron_response::RegisterVoteResponse) -> Self {
-        Self {}
-    }
-}
-
-impl From<pb::manage_neuron_response::SplitResponse>
-    for pb_api::manage_neuron_response::SplitResponse
-{
-    fn from(item: pb::manage_neuron_response::SplitResponse) -> Self {
-        Self {
-            created_neuron_id: item.created_neuron_id,
-        }
-    }
-}
-impl From<pb_api::manage_neuron_response::SplitResponse>
-    for pb::manage_neuron_response::SplitResponse
-{
-    fn from(item: pb_api::manage_neuron_response::SplitResponse) -> Self {
-        Self {
-            created_neuron_id: item.created_neuron_id,
-        }
-    }
-}
-
-impl From<pb::manage_neuron_response::MergeResponse>
-    for pb_api::manage_neuron_response::MergeResponse
-{
-    fn from(item: pb::manage_neuron_response::MergeResponse) -> Self {
-        Self {
-            source_neuron: item.source_neuron.map(|x| x.into()),
-            target_neuron: item.target_neuron.map(|x| x.into()),
-            source_neuron_info: item.source_neuron_info.map(|x| x.into()),
-            target_neuron_info: item.target_neuron_info.map(|x| x.into()),
-        }
-    }
-}
-impl From<pb_api::manage_neuron_response::MergeResponse>
-    for pb::manage_neuron_response::MergeResponse
-{
-    fn from(item: pb_api::manage_neuron_response::MergeResponse) -> Self {
-        Self {
-            source_neuron: item.source_neuron.map(|x| x.into()),
-            target_neuron: item.target_neuron.map(|x| x.into()),
-            source_neuron_info: item.source_neuron_info.map(|x| x.into()),
-            target_neuron_info: item.target_neuron_info.map(|x| x.into()),
-        }
-    }
-}
-
-impl From<pb::manage_neuron_response::DisburseToNeuronResponse>
-    for pb_api::manage_neuron_response::DisburseToNeuronResponse
-{
-    fn from(item: pb::manage_neuron_response::DisburseToNeuronResponse) -> Self {
-        Self {
-            created_neuron_id: item.created_neuron_id,
-        }
-    }
-}
-impl From<pb_api::manage_neuron_response::DisburseToNeuronResponse>
-    for pb::manage_neuron_response::DisburseToNeuronResponse
-{
-    fn from(item: pb_api::manage_neuron_response::DisburseToNeuronResponse) -> Self {
-        Self {
-            created_neuron_id: item.created_neuron_id,
-        }
-    }
-}
-
-impl From<pb::manage_neuron_response::ClaimOrRefreshResponse>
-    for pb_api::manage_neuron_response::ClaimOrRefreshResponse
-{
-    fn from(item: pb::manage_neuron_response::ClaimOrRefreshResponse) -> Self {
-        Self {
-            refreshed_neuron_id: item.refreshed_neuron_id,
-        }
-    }
-}
-impl From<pb_api::manage_neuron_response::ClaimOrRefreshResponse>
-    for pb::manage_neuron_response::ClaimOrRefreshResponse
-{
-    fn from(item: pb_api::manage_neuron_response::ClaimOrRefreshResponse) -> Self {
-        Self {
-            refreshed_neuron_id: item.refreshed_neuron_id,
-        }
-    }
-}
-
-impl From<pb::manage_neuron_response::Command> for pb_api::manage_neuron_response::Command {
-    fn from(item: pb::manage_neuron_response::Command) -> Self {
-        match item {
-            pb::manage_neuron_response::Command::Error(v) => {
-                pb_api::manage_neuron_response::Command::Error(v.into())
-            }
-            pb::manage_neuron_response::Command::Configure(v) => {
-                pb_api::manage_neuron_response::Command::Configure(v.into())
-            }
-            pb::manage_neuron_response::Command::Disburse(v) => {
-                pb_api::manage_neuron_response::Command::Disburse(v.into())
-            }
-            pb::manage_neuron_response::Command::Spawn(v) => {
-                pb_api::manage_neuron_response::Command::Spawn(v.into())
-            }
-            pb::manage_neuron_response::Command::Follow(v) => {
-                pb_api::manage_neuron_response::Command::Follow(v.into())
-            }
-            pb::manage_neuron_response::Command::MakeProposal(v) => {
-                pb_api::manage_neuron_response::Command::MakeProposal(v.into())
-            }
-            pb::manage_neuron_response::Command::RegisterVote(v) => {
-                pb_api::manage_neuron_response::Command::RegisterVote(v.into())
-            }
-            pb::manage_neuron_response::Command::Split(v) => {
-                pb_api::manage_neuron_response::Command::Split(v.into())
-            }
-            pb::manage_neuron_response::Command::DisburseToNeuron(v) => {
-                pb_api::manage_neuron_response::Command::DisburseToNeuron(v.into())
-            }
-            pb::manage_neuron_response::Command::ClaimOrRefresh(v) => {
-                pb_api::manage_neuron_response::Command::ClaimOrRefresh(v.into())
-            }
-            pb::manage_neuron_response::Command::MergeMaturity(v) => {
-                pb_api::manage_neuron_response::Command::MergeMaturity(v.into())
-            }
-            pb::manage_neuron_response::Command::Merge(v) => {
-                pb_api::manage_neuron_response::Command::Merge(v.into())
-            }
-            pb::manage_neuron_response::Command::StakeMaturity(v) => {
-                pb_api::manage_neuron_response::Command::StakeMaturity(v.into())
-            }
-            pb::manage_neuron_response::Command::RefreshVotingPower(v) => {
-                pb_api::manage_neuron_response::Command::RefreshVotingPower(v.into())
-            }
-        }
-    }
-}
-impl From<pb_api::manage_neuron_response::Command> for pb::manage_neuron_response::Command {
-    fn from(item: pb_api::manage_neuron_response::Command) -> Self {
-        match item {
-            pb_api::manage_neuron_response::Command::Error(v) => {
-                pb::manage_neuron_response::Command::Error(v.into())
-            }
-            pb_api::manage_neuron_response::Command::Configure(v) => {
-                pb::manage_neuron_response::Command::Configure(v.into())
-            }
-            pb_api::manage_neuron_response::Command::Disburse(v) => {
-                pb::manage_neuron_response::Command::Disburse(v.into())
-            }
-            pb_api::manage_neuron_response::Command::Spawn(v) => {
-                pb::manage_neuron_response::Command::Spawn(v.into())
-            }
-            pb_api::manage_neuron_response::Command::Follow(v) => {
-                pb::manage_neuron_response::Command::Follow(v.into())
-            }
-            pb_api::manage_neuron_response::Command::MakeProposal(v) => {
-                pb::manage_neuron_response::Command::MakeProposal(v.into())
-            }
-            pb_api::manage_neuron_response::Command::RegisterVote(v) => {
-                pb::manage_neuron_response::Command::RegisterVote(v.into())
-            }
-            pb_api::manage_neuron_response::Command::Split(v) => {
-                pb::manage_neuron_response::Command::Split(v.into())
-            }
-            pb_api::manage_neuron_response::Command::DisburseToNeuron(v) => {
-                pb::manage_neuron_response::Command::DisburseToNeuron(v.into())
-            }
-            pb_api::manage_neuron_response::Command::ClaimOrRefresh(v) => {
-                pb::manage_neuron_response::Command::ClaimOrRefresh(v.into())
-            }
-            pb_api::manage_neuron_response::Command::MergeMaturity(v) => {
-                pb::manage_neuron_response::Command::MergeMaturity(v.into())
-            }
-            pb_api::manage_neuron_response::Command::Merge(v) => {
-                pb::manage_neuron_response::Command::Merge(v.into())
-            }
-            pb_api::manage_neuron_response::Command::StakeMaturity(v) => {
-                pb::manage_neuron_response::Command::StakeMaturity(v.into())
-            }
-            pb_api::manage_neuron_response::Command::RefreshVotingPower(v) => {
-                pb::manage_neuron_response::Command::RefreshVotingPower(v.into())
-            }
-        }
-    }
-}
-
 impl From<pb::GovernanceError> for pb_api::GovernanceError {
     fn from(item: pb::GovernanceError) -> Self {
         Self {
