+++ rs/nns/governance/tests/merge_neurons.rs
@@ -13,9 +13,12 @@ use ic_nns_governance::{
     governance::{Environment, MAX_DISSOLVE_DELAY_SECONDS},
     pb::v1::{
         manage_neuron::{Command, Merge},
-        manage_neuron_response::{Command as CommandResponse, MergeResponse},
         ManageNeuron, NetworkEconomics,
     },
+};
+use ic_nns_governance_api::pb::v1::{
+    self as api,
+    manage_neuron_response::{Command as CommandResponse, MergeResponse},
 };
 use proptest::prelude::{proptest, TestCaseError};
 
@@ -142,6 +145,7 @@ fn do_test_merge_neurons(
                     .with_neuron(&source_neuron_id, |n| n
                         .clone()
                         .into_proto(nns.governance.voting_power_economics(), now_seconds))
+                    .map(api::Neuron::from)
                     .unwrap()
             );
             pretty_assertions::assert_eq!(
@@ -151,18 +155,21 @@ fn do_test_merge_neurons(
                     .with_neuron(&target_neuron_id, |n| n
                         .clone()
                         .into_proto(nns.governance.voting_power_economics(), now_seconds))
+                    .map(api::Neuron::from)
                     .unwrap()
             );
             pretty_assertions::assert_eq!(
                 source_neuron_info,
                 nns.governance
                     .get_neuron_info(&source_neuron_id, controller)
+                    .map(api::NeuronInfo::from)
                     .unwrap()
             );
             pretty_assertions::assert_eq!(
                 target_neuron_info,
                 nns.governance
                     .get_neuron_info(&target_neuron_id, controller)
+                    .map(api::NeuronInfo::from)
                     .unwrap()
             );
         }
